stage 1 transaction API
