+++ .rustfmt.toml
@@ -1,0 +1,3 @@
+max_width = 80
+tab_spaces = 2
+edition = "2018"
